wsd: split time and clock tests to LoggingWhiteBoxTests.cpp
